// src/grpc/gen/nockchain/public/v2/nockchain.ts
export interface WalletSendTransactionRequest {
  tx_id?: Hash | undefined;
  raw_tx?: RawTransaction | undefined;
}
// ...
export const WalletSendTransactionRequest: MessageFns<WalletSendTransactionRequest> = {
  encode(message: WalletSendTransactionRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.tx_id !== undefined) {
      Hash.encode(message.tx_id, writer.uint32(10).fork()).join();
    }
    if (message.raw_tx !== undefined) {
      RawTransaction.encode(message.raw_tx, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): WalletSendTransactionRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseWalletSendTransactionRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.tx_id = Hash.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.raw_tx = RawTransaction.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): WalletSendTransactionRequest {
    return {
      tx_id: isSet(object.txId)
        ? Hash.fromJSON(object.txId)
        : isSet(object.tx_id)
        ? Hash.fromJSON(object.tx_id)
        : undefined,
      raw_tx: isSet(object.rawTx)
        ? RawTransaction.fromJSON(object.rawTx)
        : isSet(object.raw_tx)
        ? RawTransaction.fromJSON(object.raw_tx)
        : undefined,
    };
  },

  toJSON(message: WalletSendTransactionRequest): unknown {
    const obj: any = {};
    if (message.tx_id !== undefined) {
      obj.txId = Hash.toJSON(message.tx_id);
    }
    if (message.raw_tx !== undefined) {
      obj.rawTx = RawTransaction.toJSON(message.raw_tx);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<WalletSendTransactionRequest>, I>>(base?: I): WalletSendTransactionRequest {
    return WalletSendTransactionRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<WalletSendTransactionRequest>, I>>(object: I): WalletSendTransactionRequest {
    const message = createBaseWalletSendTransactionRequest();
    message.tx_id = (object.tx_id !== undefined && object.tx_id !== null) ? Hash.fromPartial(object.tx_id) : undefined;
    message.raw_tx = (object.raw_tx !== undefined && object.raw_tx !== null)
      ? RawTransaction.fromPartial(object.raw_tx)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}